# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "The REST API a Quark device serves to its Flutter clients. Every endpoint except\n/auth/setup, /auth/login, /auth/recover, /auth/request-account and /auth/status needs a\nsession token. Sign in with POST /auth/login, then click Authorize and enter the word\nBearer, a space, and the token.",
        "title": "Quark API",
        "contact": {},
        "version": "v0"
    },
    "basePath": "/api/v0",
    "paths": {
...
        "/files": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "merges files across all managed devices for the given filePath. If deviceSerial is empty, list files across all devices. Otherwise, only for the specified device",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Lists files",
                "parameters": [
                    {
                        "type": "string",
                        "description": "File dir to list",
                        "name": "rootDir",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Device serial number to filter by",
                        "name": "serial",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/v0_files.FileNodeJSON"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Enqueue a file move operation between paths/devices. Needs write access on the file and on the folder it moves into.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Move or rename a file",
                "parameters": [
                    {
                        "description": "Move file request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v0_files.moveFileRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Ok",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Move files to the device's trash (internal storage included), returning immediately. They can be restored through /trash/restore until the hourly purge deletes them after the retention period. DB cleanup and events are dispatched in the background.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Delete files",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Root directory",
                        "name": "rootDir",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "Array of file paths to delete",
                        "name": "filePaths",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Device serial number to filter by",
                        "name": "serial",
                        "in": "query"
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    }
                }
            }
        },
...
        "/files/upload-session": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Reserve a session for one file; the bytes follow as chunks on PUT. Needs write access on the directory the file lands in. The session belongs to the caller: it is not found for anyone else. A name already in use is a 409 unless overwrite or keepBoth says what to do about it.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Open a resumable upload session",
                "parameters": [
                    {
                        "description": "File the session will carry",
                        "name": "session",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v0_files.createUploadSessionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v0_files.createUploadSessionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    }
                }
            }
        },
...
        "/jobs/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns one background job by id. A caller who is not an admin gets only a job they queued whose file they can still read, with error left blank; any other job is 404.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "Get a background job",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/jobutil.Job"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Cancels a pending or running job and returns it, now canceled. A pending job never runs; a running job is stopped and cleans up after itself. Only the account that queued the job, or an admin, may cancel it; any other caller gets 404.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "Cancel a background job",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/jobutil.Job"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    },
                    "409": {
                        "description": "Conflict — the job has already finished",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "jobutil.Job": {
            "type": "object",
            "properties": {
                "attempts": {
                    "description": "Attempts is how many times the job has started running: 0 before its\nfirst run. A retry leaves it alone; the next run bumps it.",
                    "type": "integer"
                },
                "createdAt": {
                    "description": "CreatedAt is when the job was queued.",
                    "type": "string"
                },
                "error": {
                    "description": "Error is diagnostic text for a failed job, not copy for a user.",
                    "type": "string"
                },
                "finishedAt": {
                    "description": "FinishedAt is null until the job completes, fails, or is canceled.",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "kind": {
                    "type": "string"
                },
                "lane": {
                    "description": "Lane is the concurrency lane within Kind the job runs in, such as \"copy\"\nfor a transcode that only copies streams. It is \"\" for a kind with one\nlane.",
                    "type": "string"
                },
                "name": {
                    "description": "Name is display text, such as \"Convert vacation.mkv to MOV\".",
                    "type": "string"
                },
                "params": {
                    "description": "Params is the kind-specific JSON object the job was enqueued with, so a\nclient can build its own label from Kind and Params.",
                    "type": "object"
                },
                "progress": {
                    "type": "number"
                },
                "startedAt": {
                    "description": "StartedAt is null until the job starts running.",
                    "type": "string"
                },
                "status": {
                    "enum": [
                        "pending",
                        "running",
                        "completed",
                        "failed",
                        "canceled"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/jobutil.Status"
                        }
                    ]
                }
            }
        },
        "jobutil.Status": {
            "type": "string",
            "enum": [
                "pending",
                "running",
                "completed",
                "failed",
                "canceled"
            ],
            "x-enum-varnames": [
                "StatusPending",
                "StatusRunning",
                "StatusCompleted",
                "StatusFailed",
                "StatusCanceled"
            ]
        },
...
        "serverutil.ContentType": {
            "type": "string",
            "enum": [
                "text/html",
                "application/json"
            ],
            "x-enum-varnames": [
                "ContentTypeHTML",
                "ContentTypeJSON"
            ]
        },
        "serverutil.Response": {
            "type": "object",
            "properties": {
                "contentType": {
                    "$ref": "#/definitions/serverutil.ContentType"
                },
                "data": {},
                "error": {},
                "statusCode": {
                    "type": "integer"
                }
            }
        },
...
        "v0_files.FileNodeJSON": {
            "type": "object",
            "properties": {
                "compressedSize": {
                    "type": "integer"
                },
                "deviceName": {
                    "type": "string"
                },
                "devicePath": {
                    "type": "string"
                },
                "deviceSerial": {
                    "type": "string"
                },
                "dirPath": {
                    "description": "Directory path containing the file, for easier client-side handling",
                    "type": "string"
                },
                "fileType": {
                    "description": "Kept for older clients; route viewers by file name instead",
                    "type": "string"
                },
                "fullPath": {
                    "type": "string"
                },
                "isDir": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                }
            }
        },
...
        "v0_files.createUploadSessionRequest": {
            "type": "object",
            "properties": {
                "fileName": {
                    "type": "string"
                },
                "keepBoth": {
                    "type": "boolean"
                },
                "overwrite": {
                    "type": "boolean"
                },
                "rootDir": {
                    "type": "string"
                },
                "serial": {
                    "type": "string"
                },
                "totalSize": {
                    "type": "integer"
                }
            }
        },
        "v0_files.createUploadSessionResponse": {
            "type": "object",
            "properties": {
                "expiresAt": {
                    "type": "string"
                },
                "offset": {
                    "type": "integer"
                },
                "sessionId": {
                    "type": "string"
                }
            }
        },
        "v0_files.moveFileRequest": {
            "type": "object",
            "properties": {
                "newDeviceSerial": {
                    "type": "string"
                },
                "newFilePath": {
                    "type": "string"
                },
                "oldDeviceSerial": {
                    "type": "string"
                },
                "oldFilePath": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "A session token from POST /auth/login, entered as the word Bearer, a space, and the token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}